[WIP] Replace usage of `musicbrainzngs` in `musicbrainz` plugin with `mbzero`
WIP: Update poetry.lock once new `mbzero` version has been released.

`mbzero` is a lighter interface to access the MusicBrainz API, so a
`MbInterface` has been created to re-add similar functions that
`musicbrainzngs` provided.

Create a shared `MbInterface` outside the plugin so that it can later be
reused by other plugins, using the same rate limiter.
- Instead of the `musicbrainz` plugin reading the shared config values,
  it is now `MbInterface` that does it. This makes sure these values are
  correctly handled in case the `musicbrainz` plugin is disabled, but
  not the others.

Rate limiting is not done by `mbzero`, so it has been implemented here.
The implementation is not specific to `MbInterface` and may be reused by
others if needed in the future.
- A rate-limiting bug has also been fixed compared to the
  `musicbrainzngs` implementation
  - See issue 294 on `python-musicbrainzngs`

Tests have been modified as the structure of data received is a bit
different than before.

Signed-off-by: Louis Rannou <louson@gresille.org>
Signed-off-by: Nicolas Mémeint <git@nicomem.com>