Fix JSON data containing many None values
When MusicBrainz returns data, it can do so in two formats:
- XML, which was what used the previous musicbrainzngs lib
- JSON, which is what mbzero uses

The thing is that for non-present values in the database, the two
formats handle these differently:
- XML: The keys are simply not included in the data
- JSON: The keys are included and set with the None value

Since the rest of the code worked with the XML way of doing things, and
also because it is easier to work with (e.g. "key in dict" to check if
set), I have added a cleanup step after the JSON parsing that removes
all keys with None values recursively.