Fix `mbzero` exception handling
The previous exception handling did not work correctly: for example,
trying to match by ID and giving an invalid ID would raise an exception
and stop the program.

As per the `mbzero` documentation and source code, all the requests we
make through `mbzero` either succeed, or raise a `mbzerror.MbzWebServiceError`
(more specifically a `MbzRequestError` but it does not seem public as
per the documentation, and does not brind any useful information).

The problem is that this single exception type is not fine-grained
enough for our use-cases.
I have added custom error types and a conversion step to allow for more
precise exception handling.
An upstream issue has been created for implementing something of the
sort directly in `mbzero` as we need to use things that are not
documented and thus may not have strong stability guarantees.