Clean, fix and move musicbrainz rate limiter to utils
** Clean

The rate limiter code that was previously in the `python-musicbrainzngs`
was copied locally in the first commit because `mbzero` does not include
one.

This rate limiter uses a global state (singleton) which means that it
can only be used by the musicbrainz plugin, wheras this rate-limiting
code may be useful elsewhere.

To avoid using a singleton and instead store the configuration in an
object, the rate limiter has been changed from a decorator to an object
that is used in a `with` block, similarly to a `threading.Lock`.

** Move

Since it is now not limited to the musicbrainz plugin, it has been moved
to `utils/`.

** Fix

While writing tests, I found a bug with the `musicbrainzngs`
implementation.

When we are rate-limited, we sleep until we should not be limited anymore.

The code to compute this duration was:
`(1.0 - self.remaining_requests) * (self.reqs_per_interval / self.interval_sec)`

This is wrong and should instead be:
`(1.0 - self.remaining_requests) * (self.interval_sec / self.reqs_per_interval)`

To verify this, suppose that we want 10 reqs per 0.1 seconds:
- We should wait around 0.01 second to be able to do 1 req
- `self.reqs_per_interval / self.interval_sec = 10 / 0.1 = 100`
- `self.interval_sec / self.reqs_per_interval = 0.1 / 10 = 0.01`
  - Also looking at the units, the first one returns a `req/s^2` while
    the second returns a `s`