Replace usage of `musicbrainzngs` in `listenbrainz` plugin with new `MbInterface`
This migration is quite basic.

There is one difference compared to before: when searching for
recordings, the `track_name` was simply used as a non-field query and is
now linked to the `recording` field.

This should be more precise, but if this is not wanted, the
`search_recordings` can be changed to also take a query instead of only
fields.